requirements.txt
# ============================================================================
# 瞬语 AI - Python 依赖列表
# 生成时间: 2025-12-26
# Python 版本: 3.10
# 总大小: ~1.77 GB
# ============================================================================

# --------------------------------------------------------------------------
# 核心 ASR 依赖
# --------------------------------------------------------------------------
# ONNX Runtime (用于 SenseVoice ONNX 模型)
onnxruntime==1.23.2

# FunASR / Paraformer (备选模型)
funasr==1.2.9
funasr-onnx==0.4.1

# PyTorch (funasr/paraformer 需要)
torch==2.5.1+cpu
torchaudio==2.5.1+cpu

# --------------------------------------------------------------------------
# 语音处理管线
# --------------------------------------------------------------------------
# VAD (语音活动检测)
silero-vad==6.2.0

# 音频处理
soundfile==0.12.1
librosa==0.11.0
pydub==0.25.1
soxr==1.0.0
audioread==3.1.0

# --------------------------------------------------------------------------
# 模型相关
# --------------------------------------------------------------------------
huggingface_hub==1.2.3
modelscope==1.33.0
sentencepiece==0.2.0

# --------------------------------------------------------------------------
# 科学计算
# --------------------------------------------------------------------------
numpy==1.24.4
scipy==1.15.3
scikit-learn==1.7.2
numba==0.63.1

# --------------------------------------------------------------------------
# 工具库
# --------------------------------------------------------------------------
PyYAML==6.0.3
requests==2.32.5
tqdm==4.67.1
coloredlogs==15.0.1
protobuf==6.33.2

# --------------------------------------------------------------------------
# 阿里云 OSS (modelscope 依赖)
# --------------------------------------------------------------------------
aliyun-python-sdk-core==2.16.0
oss2==2.19.1

# ============================================================================
# 安装命令:
# pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cpu
# ============================================================================
